requirements.txt
flask
flask-mysql
flask-cors
python-dotenv
flask-mysqldb
selenium